[CHANGE] Moved Overlay System to own py [FIX] In some cases it not open the overlay [REMOVE] Color Mode Support
